chore: move `eligibility_index` to be under `eligibility` app
