+++ benefits/core/context/__init__.py
@@ -1,4 +1,4 @@
-from .agency import AgencySlug, agency_index, eligibility_index
+from .agency import AgencySlug, agency_index
 from .flow import SystemName
 
-__all__ = ["AgencySlug", "SystemName", "agency_index", "eligibility_index"]
+__all__ = ["AgencySlug", "SystemName", "agency_index"]
