+++ benefits/eligibility/context/__init__.py
@@ -1,0 +1,3 @@
+from .agency import eligibility_index
+
+__all__ = ["eligibility_index"]
