+++ benefits/locale/en/LC_MESSAGES/django.po
@@ -6,7 +6,7 @@
 msgid ""
 msgstr ""
 "Report-Msgid-Bugs-To: https://github.com/cal-itp/benefits/issues \n"
-"POT-Creation-Date: 2025-02-26 20:37-0800\n"
+"POT-Creation-Date: 2025-03-06 15:05-0800\n"
 "Language: English\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -28,43 +28,6 @@ msgid "Get a reduced fare on SacRT buses when you tap to ride"
 msgstr ""
 
 msgid "Get a reduced fare on Santa Barbara MTD buses when you tap to ride"
-msgstr ""
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All CST transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All MST transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All Nevada County Connects "
-"transit benefits reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All SacRT transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-
-msgid ""
-"With the new Tap2Ride fare system launching in 2025, SacRT riders that are "
-"eligible for discount fares can enroll below. The discount benefit will be "
-"available on all SacRT buses beginning in early 2025 and will include light "
-"rail tap devices later in the year."
-msgstr ""
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All SBMTD transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
 msgstr ""
 
 msgid ""
@@ -404,6 +367,43 @@ msgid "You have successfully logged out. Thank you for using Cal-ITP Benefits!"
 msgstr ""
 
 msgid "Which transit benefit would you like to enroll in?"
+msgstr ""
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All CST transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All MST transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All Nevada County Connects "
+"transit benefits reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All SacRT transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+
+msgid ""
+"With the new Tap2Ride fare system launching in 2025, SacRT riders that are "
+"eligible for discount fares can enroll below. The discount benefit will be "
+"available on all SacRT buses beginning in early 2025 and will include light "
+"rail tap devices later in the year."
+msgstr ""
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All SBMTD transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
 msgstr ""
 
 msgid "Choose this benefit"
