+++ benefits/locale/es/LC_MESSAGES/django.po
@@ -6,7 +6,7 @@
 msgid ""
 msgstr ""
 "Report-Msgid-Bugs-To: https://github.com/cal-itp/benefits/issues \n"
-"POT-Creation-Date: 2025-02-26 20:37-0800\n"
+"POT-Creation-Date: 2025-03-06 15:05-0800\n"
 "Language: Español\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -39,59 +39,6 @@ msgid "Get a reduced fare on Santa Barbara MTD buses when you tap to ride"
 msgstr ""
 "Obtenga una tarifa reducida en los autobuses del Santa Barbara MTD cuando "
 "toque para viajar"
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All CST transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-"Cal-ITP no almacena su información. Todos los beneficios de tránsito de CST "
-"reducen las tarifas de autobuses en un 50%%."
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All MST transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-"Cal-ITP no almacena su información. Todos los beneficios de tránsito de MST "
-"reducen las tarifas de autobuses en un 50%%."
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All Nevada County Connects "
-"transit benefits reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-"Cal-ITP no almacena su información. Todos los beneficios de tránsito de "
-"Nevada County Connects reducen las tarifas de autobuses en un 50%%."
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All SacRT transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-"Cal-ITP no almacena su información. Todos los beneficios de tránsito de "
-"SacRT reducen las tarifas en un 50%% para el servicio de autobús en rutas "
-"fijas."
-
-msgid ""
-"With the new Tap2Ride fare system launching in 2025, SacRT riders that are "
-"eligible for discount fares can enroll below. The discount benefit will be "
-"available on all SacRT buses beginning in early 2025 and will include light "
-"rail tap devices later in the year."
-msgstr ""
-"Con el nuevo sistema de tarifas Tap2Ride que se lanzará en 2025, los "
-"pasajeros de SacRT que sean elegibles para tarifas con descuento pueden "
-"inscribirse a continuación. El beneficio de descuento estará disponible en "
-"todos los autobuses de SacRT a partir de principios de 2025 e incluirá "
-"lectores de tarjeta en el tren ligero más adelante en el año."
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All SBMTD transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-"Cal-ITP no almacena su información. Todos los beneficios de tránsito de "
-"SBMTD le ofrecen a los pasajeros tarifas de autobús reducidas."
 
 msgid ""
 "The Benefits tool is provided by California Integrated Travel Project (Cal-"
@@ -554,6 +501,59 @@ msgstr "Ha cerrado sesión correctamente. ¡Gracias por usar Cal-ITP Benefits!"
 
 msgid "Which transit benefit would you like to enroll in?"
 msgstr "¿A qué beneficio de tránsito le gustaría inscribirse?"
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All CST transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+"Cal-ITP no almacena su información. Todos los beneficios de tránsito de CST "
+"reducen las tarifas de autobuses en un 50%%."
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All MST transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+"Cal-ITP no almacena su información. Todos los beneficios de tránsito de MST "
+"reducen las tarifas de autobuses en un 50%%."
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All Nevada County Connects "
+"transit benefits reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+"Cal-ITP no almacena su información. Todos los beneficios de tránsito de "
+"Nevada County Connects reducen las tarifas de autobuses en un 50%%."
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All SacRT transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+"Cal-ITP no almacena su información. Todos los beneficios de tránsito de "
+"SacRT reducen las tarifas en un 50%% para el servicio de autobús en rutas "
+"fijas."
+
+msgid ""
+"With the new Tap2Ride fare system launching in 2025, SacRT riders that are "
+"eligible for discount fares can enroll below. The discount benefit will be "
+"available on all SacRT buses beginning in early 2025 and will include light "
+"rail tap devices later in the year."
+msgstr ""
+"Con el nuevo sistema de tarifas Tap2Ride que se lanzará en 2025, los "
+"pasajeros de SacRT que sean elegibles para tarifas con descuento pueden "
+"inscribirse a continuación. El beneficio de descuento estará disponible en "
+"todos los autobuses de SacRT a partir de principios de 2025 e incluirá "
+"lectores de tarjeta en el tren ligero más adelante en el año."
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All SBMTD transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+"Cal-ITP no almacena su información. Todos los beneficios de tránsito de "
+"SBMTD le ofrecen a los pasajeros tarifas de autobús reducidas."
 
 msgid "Choose this benefit"
 msgstr "Elegir este beneficio"
